+++ extension.js
@@ -4,6 +4,7 @@ const vscode = require('vscode');
 const fs = require('fs');
 const path = require("path");
 
+const mHook = require("./commands/mHook");
 const mTemplate = require("./commands/mTemplate");
 const mMigrationAddTable = require("./commands/mMigrationAddTable");
 const mMigrationRenameTable = require("./commands/mMigrationRenameTable");
@@ -34,6 +35,9 @@ function activate(context) {
 	// Now provide the implementation of the command with  registerCommand
 	// The commandId parameter must match the command field in package.json
 
+	let hook = vscode.commands.registerCommand('make.ci3_hook', function () {
+		mHook(vscode, fs, path, pathwork);
+	});
 	let template = vscode.commands.registerCommand('make.ci3_template', function () {
 		mTemplate(vscode, fs, path, pathwork);
 	});
@@ -83,6 +87,7 @@ function activate(context) {
 	});
 
 	
+	context.subscriptions.push(hook);
 	context.subscriptions.push(template);
 	context.subscriptions.push(migration_add_table);
 	context.subscriptions.push(migration_add_column);